requirements.txt
streamlit
nltk
sentence-transformers
scikit-learn
PyMuPDF
python-docx